refactor(changelog): move source under gitlab folder
