+++ lib/workers/repository/update/pr/changelog/api.ts
@@ -1,7 +1,7 @@
 import { BitbucketChangeLogSource } from './bitbucket/source';
+import { GitLabChangeLogSource } from './gitlab/source';
 import type { ChangeLogSource } from './source';
 import { GitHubChangeLogSource } from './source-github';
-import { GitLabChangeLogSource } from './source-gitlab';
 
 const api = new Map<string, ChangeLogSource>();
 export default api;
