+++ lib/workers/repository/update/pr/changelog/gitlab/source.ts
@@ -1,5 +1,5 @@
-import type { BranchUpgradeConfig } from '../../../../types';
-import { ChangeLogSource } from './source';
+import type { BranchUpgradeConfig } from '../../../../../types';
+import { ChangeLogSource } from '../source';
 
 export class GitLabChangeLogSource extends ChangeLogSource {
   constructor() {
